Protected all API endpoints with Sanctum

+++ routes/api.php
@@ -2,6 +2,7 @@
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
+use App\Http\Controllers\AuthController;
 use App\Http\Controllers\TaskController;
 use App\Http\Controllers\ProjectController;
 
@@ -16,40 +17,42 @@ use App\Http\Controllers\ProjectController;
 |
 */
 
+Route::post('/register/', [AuthController::class, 'createUser']);
+
 Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
 
-Route::get('/tasks/', [TaskController::class, 'getTasks']);
+Route::get('/tasks/', [TaskController::class, 'getTasks'])->middleware('auth:sanctum');
 
-Route::get('/tasks/inbox', [TaskController::class, 'getInboxTasks']);
+Route::get('/tasks/inbox', [TaskController::class, 'getInboxTasks'])->middleware('auth:sanctum');
 
-Route::get('/tasks/today', [TaskController::class, 'getTodayTasks']);
+Route::get('/tasks/today', [TaskController::class, 'getTodayTasks'])->middleware('auth:sanctum');
 
-Route::post('/tasks/add', [TaskController::class, 'addTask']);
+Route::post('/tasks/add', [TaskController::class, 'addTask'])->middleware('auth:sanctum');
 
-Route::post('/tasks/add-task-to-today', [TaskController::class, 'addTaskToToday']);
+Route::post('/tasks/add-task-to-today', [TaskController::class, 'addTaskToToday'])->middleware('auth:sanctum');
 
-Route::put('/tasks/{task}/toggle-complete', [TaskController::class, 'toggleComplete']);
+Route::put('/tasks/{task}/toggle-complete', [TaskController::class, 'toggleComplete'])->middleware('auth:sanctum');
 
-Route::get('/tasks/{task}', [TaskController::class, 'show']);
+Route::get('/tasks/{task}', [TaskController::class, 'show'])->middleware('auth:sanctum');
 
-Route::put('/tasks/{task}/edit', [TaskController::class, 'edit']);
+Route::put('/tasks/{task}/edit', [TaskController::class, 'edit'])->middleware('auth:sanctum');
 
-Route::put('/tasks/{task}/schedule-on', [TaskController::class, 'scheduleOn']);
+Route::put('/tasks/{task}/schedule-on', [TaskController::class, 'scheduleOn'])->middleware('auth:sanctum');
 
-Route::delete('/tasks/{task}/delete', [TaskController::class, 'delete']);
+Route::delete('/tasks/{task}/delete', [TaskController::class, 'delete'])->middleware('auth:sanctum');
 
-Route::get('/projects/', [ProjectController::class, 'getProjects']);
+Route::get('/projects/', [ProjectController::class, 'getProjects'])->middleware('auth:sanctum');
 
-Route::post('/projects/add', [ProjectController::class, 'create']);
+Route::post('/projects/add', [ProjectController::class, 'create'])->middleware('auth:sanctum');
 
-Route::get('/projects/{project}', [ProjectController::class, 'show']);
+Route::get('/projects/{project}', [ProjectController::class, 'show'])->middleware('auth:sanctum');
 
-Route::put('/tasks/{task}/move-to-project/{project}', [ProjectController::class, 'moveTaskToProject']);
+Route::put('/tasks/{task}/move-to-project/{project}', [ProjectController::class, 'moveTaskToProject'])->middleware('auth:sanctum');
 
-Route::post('/projects/{project}/add-task-to-project', [ProjectController::class, 'addTaskToProject']);
+Route::post('/projects/{project}/add-task-to-project', [ProjectController::class, 'addTaskToProject'])->middleware('auth:sanctum');
 
-Route::put('/projects/{project}/edit', [ProjectController::class, 'edit']);
+Route::put('/projects/{project}/edit', [ProjectController::class, 'edit'])->middleware('auth:sanctum');
 
-Route::put('/tasks/{task}/toggle-today', [TaskController::class, 'toggleToday']);
+Route::put('/tasks/{task}/toggle-today', [TaskController::class, 'toggleToday'])->middleware('auth:sanctum')->middleware('auth:sanctum');
